IR_Project_BE/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==22.2.0
avro==1.11.1
beautifulsoup4==4.12.1
bidict==0.22.1
blis==0.7.9
branca==0.6.0
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cmake==3.26.3
confection==0.0.4
confluent-kafka==2.1.0
cssselect==1.2.0
cymem==2.0.7
emoji==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.10.7
Fiona==1.9.2
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.3
flatbuffers==23.3.3
folium==0.14.0
gast==0.4.0
geographiclib==2.0
geopandas==0.12.2
geopy==2.3.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
huggingface-hub==0.13.4
idna==3.4
itsdangerous==2.1.2
jax==0.4.8
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
kafka-python==2.0.2
keras==2.12.0
langcodes==3.3.0
libclang==16.0.0
lit==16.0.1
locationtagger==0.0.1
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.0.4
mpmath==1.3.0
munch==2.5.0
murmurhash==1.0.9
networkx==3.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pydantic==1.10.7
pyproj==3.5.0
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.4.0
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
sgmllib3k==1.0.0
shapely==2.0.1
simple-websocket==0.10.0
six==1.16.0
smart-open==6.3.0
snscrape==0.6.2.20230320
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
thinc==8.1.9
tinysegmenter==0.3
tldextract==3.4.0
tokenizers==0.13.3
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.4
triton==2.0.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wasabi==1.1.1
Werkzeug==2.2.3
wrapt==1.14.1
wsproto==1.2.0
